pyproject.toml

[tool.poetry]
name = "library_of_babel"
version = "0.1.0"
description = ""
authors = ["Adam Pogwizd <adampog77@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
pycryptodome = "^3.12.0"
Flask-Cors = "^3.0.10"
gunicorn = "^20.1.0"
pillow = '^9'
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
